Move graph post order traversal from utils to trace
