+++ lib/compile.js
@@ -6,7 +6,7 @@ var url = require('url');
 var createHash = require('crypto').createHash;
 var template = require('es6-template-strings');
 var getAlias = require('./utils').getAlias;
-var getTreeModulesPostOrder = require('./utils').getTreeModulesPostOrder;
+var getTreeModulesPostOrder = require('./trace').getTreeModulesPostOrder;
 
 var compilerMap = {
   'amd': '../compilers/amd',
