+++ lib/trace.js
@@ -4,6 +4,8 @@ var toFileURL = require('./utils').toFileURL;
 var fromFileURL = require('./utils').fromFileURL;
 var asp = require('rsvp').denodeify;
 var fs = require('fs');
+var Graph = require('algorithms/data_structures/graph');
+var depthFirst = require('algorithms/graph').depthFirstSearch;
 
 module.exports = Trace;
 
@@ -399,4 +401,62 @@ Trace.getLoadDependencies = function(load, traceAllConditionals, conditionalEnv)
   return load.deps.map(function(dep) {
     return load.depMap[dep];
   });
+};
+
+function getGraphEntryPoints(graph, entryPoints) {
+  entryPoints = [].concat(entryPoints || []);
+
+  var modules = Object.keys(graph.adjList);
+  var discarded = {};
+
+  modules.forEach(function (moduleName) {
+    Object.keys(graph.adjList[moduleName]).forEach(function (depName) {
+      discarded[depName] = true;
+    });
+  });
+
+  modules.filter(function (moduleName) {
+    return !discarded[moduleName];
+  }).sort().forEach(function (moduleName) {
+    if (entryPoints.indexOf(moduleName) === -1) {
+      entryPoints.push(moduleName);
+    }
+  });
+
+  return entryPoints;
+}
+
+Trace.getTreeModulesPostOrder = function getTreeModulesPostOrder(tree, entryPoints) {
+  entryPoints = entryPoints || [];
+
+  // Post order traversal sorted module list
+  var postOrder = [];
+  var graph = new Graph(true);
+
+  // Seed graph with all relations
+  Object.keys(tree).forEach(function (moduleName) {
+    var load = tree[moduleName];
+
+    if (!graph.adjList[moduleName]) {
+      graph.addVertex(moduleName);
+    }
+
+    Trace.getLoadDependencies(load).forEach(function (depName) {
+      graph.addEdge(moduleName, depName);
+    });
+  });
+
+  // Post order traversal of graph, one per entryPoint
+  getGraphEntryPoints(graph, entryPoints).forEach(function (entryPoint) {
+    depthFirst(graph, entryPoint, {
+      leaveVertex: function (moduleName) {
+        // Avoid duplicates and modules that have been skipped by tracer
+        if (postOrder.indexOf(moduleName) === -1 && moduleName in tree) {
+          postOrder.push(moduleName);
+        }
+      }
+    });
+  });
+
+  return postOrder;
 };
