+++ lib/utils.js
@@ -1,9 +1,5 @@
 var path = require('path');
 var url = require('url');
-var trace = require('./trace');
-
-var Graph = require('algorithms/data_structures/graph');
-var depthFirst = require('algorithms/graph').depthFirstSearch;
 
 function fromFileURL(url) {
   return url.substr(7 + !!process.platform.match(/^win/)).replace(/\//g, path.sep);
@@ -179,61 +175,3 @@ function getAlias(loader, canonicalName) {
 
   return bestAlias || canonicalName;
 }
-
-function getGraphEntryPoints(graph, entryPoints) {
-  entryPoints = [].concat(entryPoints || []);
-
-  var modules = Object.keys(graph.adjList);
-  var discarded = {};
-
-  modules.forEach(function (moduleName) {
-    Object.keys(graph.adjList[moduleName]).forEach(function (depName) {
-      discarded[depName] = true;
-    });
-  });
-
-  modules.filter(function (moduleName) {
-    return !discarded[moduleName];
-  }).sort().forEach(function (moduleName) {
-    if (entryPoints.indexOf(moduleName) === -1) {
-      entryPoints.push(moduleName);
-    }
-  });
-
-  return entryPoints;
-}
-
-exports.getTreeModulesPostOrder = function getTreeModulesPostOrder(tree, entryPoints) {
-  entryPoints = entryPoints || [];
-
-  // Post order traversal sorted module list
-  var postOrder = [];
-  var graph = new Graph(true);
-
-  // Seed graph with all relations
-  Object.keys(tree).forEach(function (moduleName) {
-    var load = tree[moduleName];
-
-    if (!graph.adjList[moduleName]) {
-      graph.addVertex(moduleName);
-    }
-
-    trace.getLoadDependencies(load).forEach(function (depName) {
-      graph.addEdge(moduleName, depName);
-    });
-  });
-
-  // Post order traversal of graph, one per entryPoint
-  getGraphEntryPoints(graph, entryPoints).forEach(function (entryPoint) {
-    depthFirst(graph, entryPoint, {
-      leaveVertex: function (moduleName) {
-        // Avoid duplicates and modules that have been skipped by tracer
-        if (postOrder.indexOf(moduleName) === -1 && moduleName in tree) {
-          postOrder.push(moduleName);
-        }
-      }
-    });
-  });
-
-  return postOrder;
-};
